Refactor ChatGPT Codex Backend and Transport Layer
- Updated the ChatGPTCodexBackend to implement a protocol for better type safety and clarity.
- Introduced a new module `chatgpt_codex_transport.py` to handle WebSocket connections and related error handling.
- Removed unused WebSocket error handling code from the backend, centralizing it in the new transport module.
- Updated tests to reflect changes in the backend and transport layer, ensuring compatibility with the new structure.
- Revised TODO list to reflect current tasks and priorities.
