+++ src/pbi_agent/providers/openai_provider.py
@@ -31,9 +31,10 @@ from pbi_agent.providers import retry as provider_retry
 from pbi_agent.providers.base import Provider
 from pbi_agent.providers.chatgpt_codex_backend import (
     ChatGPTCodexBackend,
-    ChatGPTCodexWebSocketError,
+    ChatGPTCodexBackendProtocol,
     ResponsesRequestOptions,
 )
+from pbi_agent.providers.chatgpt_codex_transport import ChatGPTCodexWebSocketError
 from pbi_agent.providers.protocols.openai_responses import (
     response_history_item_for_input,
     responses_include,
@@ -107,7 +108,7 @@ class OpenAIProvider(Provider):
         self._tool_catalog = tool_catalog or ToolCatalog.from_builtin_registry()
         self._excluded_tools = default_excluded_tool_names(excluded_tools)
         self._tools: list[dict[str, Any]] = []
-        self._chatgpt_backend = ChatGPTCodexBackend(
+        self._chatgpt_backend: ChatGPTCodexBackendProtocol = ChatGPTCodexBackend(
             responses_url=self._settings.responses_url
         )
         self.refresh_tools()
